requirements.txt
torch>=2.0.0
scanpy>=1.9.0
matplotlib>=3.5.0
umap-learn>=0.5.0
scikit-learn>=1.0.0
tqdm>=4.65.0
pandas>=1.5.0
numpy>=1.23.0
